requirements.txt
streamlit
pandas
python-dotenv 
sqlalchemy
plotly
sqlalchemy-bigquery
sqlalchemy-bigquery[bqstorage]
